+++ flutter/wtc/lib/pages/nav_bar/map_pages/map.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:latlong2/latlong.dart';
-import 'package:wtc/widgets/map_card.dart';
-import 'package:wtc/widgets/map_pin.dart';
-import 'package:wtc/widgets/organization_list.dart';
+import 'package:wtc/widgets/map_widgets/map_card.dart';
+import 'package:wtc/widgets/map_widgets/map_pin.dart';
+import 'package:wtc/widgets/list_widgets/organization_list.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 MapController mapController = MapController();
