+++ flutter/wtc/lib/widgets/list_widgets/alerts_list.dart
@@ -1,22 +1,22 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_guid/flutter_guid.dart';
-import 'package:wtc/widgets/job_post/job_post.dart';
+import 'package:wtc/widgets/event_widgets/event.dart';
+import 'package:wtc/widgets/post_widgets/job_post/job_post.dart';
+import 'package:wtc/widgets/post_widgets/post.dart';
 
-// ignore: must_be_immutable
-class VolunteerPostList extends StatelessWidget {
+class AlertsList extends StatelessWidget {
   final FirebaseFirestore _firestore = FirebaseFirestore.instance;
 
-  VolunteerPostList({super.key});
+  AlertsList({super.key});
 
   @override
   Widget build(BuildContext context) {
-
     return StreamBuilder(
         stream: _firestore
             .collection('_posts')
-            .where('type', isEqualTo: 'Volunteer')
             .orderBy('timestamp', descending: true)
+            .where('type', isEqualTo: 'Alert')
             .snapshots(),
         builder: (context, AsyncSnapshot<QuerySnapshot> snapshot) {
           if (snapshot.connectionState == ConnectionState.waiting) {
@@ -32,14 +32,17 @@ class VolunteerPostList extends StatelessWidget {
             padding: const EdgeInsets.all(8),
             itemCount: snapshot.data?.docs.length ?? 0,
             itemBuilder: (BuildContext context, int index) {
-            var document = snapshot.data?.docs[index];
+              var document = snapshot.data?.docs[index];
+              String dateCreated = document?['createdAt'] as String;
+    
+              var tempTags = document?['tags'] as List<dynamic>;
+              List<String> postTags = [];
 
-            String dateCreated = document?['createdAt'] as String;
-
-            return JobPost(
-                  title: document?['title'] as String,
-                  body: document?['body'] as String,
-                  tags: const ["Volunteer"],
+              for (int i = 0; i < tempTags.length; i++) {
+                postTags.add(tempTags[i]);
+              }
+              return Post(
+                  postId: Guid(document?['postID'] as String),
                   header: document?['header'] as String,
                   userEmail: document?['user'] as String,
                   interestCount: document?['interestCount'] as int,
@@ -47,10 +50,12 @@ class VolunteerPostList extends StatelessWidget {
                       int.parse(dateCreated.split("-")[0]),
                       int.parse(dateCreated.split("-")[1]),
                       int.parse(dateCreated.split("-")[2])),
-                  postId: Guid(document?['postID'] as String),
-                  volunteer: true,
+                  title: document?['title'] as String,
+                  tags: postTags,
+                  body: document?['body'] as String,
                 );
-            },
+              }
+            ,
             separatorBuilder: (BuildContext context, int index) =>
                 const Divider(),
           );
