+++ flutter/wtc/lib/widgets/list_widgets/attending_list.dart
