+++ flutter/wtc/lib/widgets/post_widgets/job_post/job_wage_box.dart
